Fixed deep copy of Xml tags.

+++ src/common/Text/GUnicode.cpp
@@ -7,11 +7,11 @@ wchar_t *Utf8ToWide(const char *In, ssize_t InLen)
 		return NULL;
 
 	#ifdef WIN32
-		auto r = MultiByteToWideChar(CP_UTF8, 0, In, InLen, NULL, 0);
+		auto r = MultiByteToWideChar(CP_UTF8, 0, In, (int)InLen, NULL, 0);
 		if (r <= 0) return NULL;
 		auto s = new wchar_t[r + 1];
 		if (!s) return NULL;
-		r = MultiByteToWideChar(CP_UTF8, 0, In, InLen, s, r);
+		r = MultiByteToWideChar(CP_UTF8, 0, In, (int)InLen, s, r);
 		if (r <= 0) return NULL;
 		s[r] = 0;
 		return s;	
@@ -26,12 +26,12 @@ char *WideToUtf8(const wchar_t *In, ptrdiff_t InLen)
 		return NULL;
 
 	#ifdef WIN32
-		auto r = WideCharToMultiByte(CP_UTF8, 0, In, InLen, NULL, 0, 0, NULL);
+		auto r = WideCharToMultiByte(CP_UTF8, 0, In, (int)InLen, NULL, 0, 0, NULL);
 		if (r <= 0) return NULL;
 		auto s = new char[r + 1];
 		// LgiStackTrace("WideToUtf8 %p\n", s);
 		if (!s) return NULL;
-		r = WideCharToMultiByte(CP_UTF8, 0, In, InLen, s, r, 0, NULL);
+		r = WideCharToMultiByte(CP_UTF8, 0, In, (int)InLen, s, r, 0, NULL);
 		if (r <= 0) return NULL;
 		s[r] = 0;
 		return s;
